requirements.txt
openai>=1.12.0
faiss-cpu==1.7.4
python-dotenv==1.0.1
numpy==1.26.4
click==8.1.7
requests==2.31.0 